# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Alexandryn system API
  version: 0.0.0
  description: >-
    The contract between the Go server and every client (the Electron
    renderer and any LAN browser). Errors share one shape across every
    endpoint.
servers:
  - url: /
    description: The Go server's own origin (loopback by default, constitution §6).

# Path-based versioning. Health lives at the root, outside the versioned
# surface, because it predates and is independent of it.
paths:
...
  /api/v1/collections:
    get:
      operationId: listCollections
      summary: List all collections (unpaginated)
      description: >-
        Returns every collection with its work count. No pagination —
        household-scale collection count assumed manageable.
      security: []
      responses:
        '200':
          description: All collections.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [collections]
                properties:
                  collections:
                    type: array
                    items:
                      $ref: '#/components/schemas/CollectionSummary'
              example:
                collections:
                  - id: "01JXXXXXXXXXXXXXXXXXXXXXXZ"
                    name: "Classics"
                    workCount: 1
    post:
      operationId: createCollection
      summary: Create a new collection
      description: >-
        Creates a new collection. Name is validated
        (max 100 chars, no control characters).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreate'
      responses:
        '201':
          description: The created collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetail'
              example:
                id: "01JXXXXXXXXXXXXXXXXXXXXXXZ"
                name: "Classics"
                works: []
        '400':
          description: Invalid collection name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: invalid_input
                message: "name: exceeds maximum length of 100 characters"
                correlationId: 00000000-0000-0000-0000-000000000000
  /api/v1/collections/{id}:
    get:
      operationId: getCollection
      summary: Get a collection's detail including member works
      description: >-
        Returns one collection and its member Works, each with the
        per-membership added-at timestamp.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: CollectionID. Malformed → 400; not found → 404.
      responses:
        '200':
          description: The collection detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetail'
              example:
                id: "01JXXXXXXXXXXXXXXXXXXXXXXZ"
                name: "Classics"
                works:
                  - id: "01JXXXXXXXXXXXXXXXXXXXXXXX"
                    title: "Middlemarch"
                    subtitle: "A Study of Provincial Life"
                    authors:
                      - George Eliot
                    isOwned: true
                    collections:
                      - id: "01JXXXXXXXXXXXXXXXXXXXXXXZ"
                        name: "Classics"
                        addedAt: "2026-01-20T08:00:00Z"
                    addedAt: "2026-01-15T10:00:00Z"

        '400':
          description: Malformed CollectionID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: invalid_input
                message: "id: not a valid collection ID"
                correlationId: 00000000-0000-0000-0000-000000000000
        '404':
          description: No collection with that ID exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: not_found
                message: "no collection with that id"
                correlationId: 00000000-0000-0000-0000-000000000000
    patch:
      operationId: renameCollection
      summary: Rename a collection
      description: >-
        Renames a collection. Membership changes go through
        POST/DELETE /api/v1/collections/{id}/works, never bundled here.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionRename'
      responses:
        '200':
          description: The renamed collection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionDetail'
              example:
                id: "01JXXXXXXXXXXXXXXXXXXXXXXZ"
                name: "Renamed Collection"
                works: []
        '400':
          description: Invalid name or malformed CollectionID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: invalid_input
                message: "name: contains control characters"
                correlationId: 00000000-0000-0000-0000-000000000000
        '404':
          description: No collection with that ID exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: not_found
                message: "no collection with that id"
                correlationId: 00000000-0000-0000-0000-000000000000
    delete:
      operationId: deleteCollection
      summary: Delete a collection
      description: >-
        Deletes the collection and its membership rows. Non-cascading —
        Works, Editions, and LibraryEntries are untouched.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted successfully. No body.
        '400':
          description: Malformed CollectionID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: invalid_input
                message: "id: not a valid collection ID"
                correlationId: 00000000-0000-0000-0000-000000000000
        '404':
          description: No collection with that ID exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: not_found
                message: "no collection with that id"
                correlationId: 00000000-0000-0000-0000-000000000000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, correlationId]
      description: >-
        The one error shape every endpoint uses. `correlationId` matches
        the id in the server's structured logs for the same request.
      properties:
        code:
          type: string
          description: Machine-readable category (domain.Category).
          example: not_found
        message:
          type: string
          description: >-
            Human-readable, specific, no apology, says what to do next
            (constitution §11).
          example: no such endpoint
        correlationId:
          type: string
          description: Ties this response to one server log line.
          example: 00000000-0000-0000-0000-000000000000

    # ── Library domain schemas ───────────────────────────────────────
    WorkSummary:
      type: object
      additionalProperties: false
      required: [id, title, subtitle, authors, isOwned, collections]
      description: >-
        One work in the /api/v1/library paginated list.
      properties:
        id:
          type: string
          description: WorkID (internal).
        title:
          type: string
        subtitle:
          type: string
          description: Empty string when absent — never null.
        authors:
          type: array
          items:
            type: string
          description: Author display names.
        isOwned:
          type: boolean
          description: >-
            True when at least one Edition of this Work has a LibraryEntry
            (computed, not stored).
        collections:
          type: array
          items:
            $ref: '#/components/schemas/CollectionRef'
          description: Collections this Work is a member of.
        addedAt:
          type: string
          format: date-time
          description: >-
            Sort key for the added_at sort mode. Present when the work is
            owned (LibraryEntry.added_at) or wanted (most recent
            collection membership added_at). Absent only when a work is
            somehow reachable with neither (not expected in practice).
          nullable: true
...
    CollectionRef:
      type: object
      additionalProperties: false
      required: [id, name]
      description: >-
        Minimal collection reference embedded in WorkSummary and
        WorkDetail.
      properties:
        id:
          type: string
          description: CollectionID (internal).
        name:
          type: string
        addedAt:
          type: string
          format: date-time
          description: >-
            Per-membership added-at timestamp — when this Work was added
            to this Collection.
...
    CollectionSummary:
      type: object
      additionalProperties: false
      required: [id, name, workCount]
      description: >-
        One collection in the GET /api/v1/collections list.
      properties:
        id:
          type: string
        name:
          type: string
        workCount:
          type: integer
          description: Number of Works currently in this Collection.
    CollectionDetail:
      type: object
      additionalProperties: false
      required: [id, name, works]
      description: >-
        Response body for GET /api/v1/collections/{id}.
      properties:
        id:
          type: string
        name:
          type: string
        works:
          type: array
          items:
            $ref: '#/components/schemas/WorkSummary'
          description: >-
            Works in this Collection, each carrying the per-membership
            added-at timestamp.
    CollectionCreate:
      type: object
      additionalProperties: false
      required: [name]
      description: >-
        Request body for POST /api/v1/collections.
      properties:
        name:
          type: string
          maxLength: 100
          description: Collection name.
    CollectionRename:
      type: object
      additionalProperties: false
      required: [name]
      description: >-
        Request body for PATCH /api/v1/collections/{id}.
      properties:
        name:
          type: string
          maxLength: 100
...
security:
  - bearerAuth: []